refactor(reviewer-assign): re-home scanner inline tests to sibling tests.rs (t-…-17 slice 5)
The t-…-17 additions to pr_state/scanner.rs (record_terminal wire + B3 retain-on-error + t21/B3 tests) pushed it from t-9's 2338 to 2564 LOC, over the 2500-LOC anti-monolith ceiling (tests/src_file_size_invariant.rs). Per that gate's prescribed fix + the established pattern (scanner/review_repro_daemon_ci_pr.rs; pr_state/mod.rs de-grandfather), re-home the inline #[cfg(test)] mod tests into sibling src/daemon/pr_state/scanner/tests.rs (is_test_file exempt). Pure move — super:: paths unchanged; 3 include_str!("scanner.rs") → "../scanner.rs" for the deeper file. scanner.rs now 1073 LOC; all 33 scanner tests pass; gate green.

Agend-Agent: archfix-opus-2
Agend-Task: t-20260713002749179875-15764-17
Agend-Branch: fix/reviewer-assignment-outbox
Agend-Issued-At: 2026-07-13T00:35:43.419484+00:00
